Modularizing core content
