+++ src/core.js
@@ -1,0 +1,7 @@
+module.exports = {
+	File: require('./file'),
+	HttpRequest: require('./httpRequest'),
+	Logger: require('./log/logger'),
+	LoggerFactory: require('./log/loggerFactory'),
+	Router: require('./router')
+}
